Update orm.md
I reminded users that they can auto generate their tables by including the orm.RunSyncdb command in their main.go file. This was seen in the wetalk project and was a point of confusion prior to seeing it used in wetalk. It should help clarify how to autogenerate tables with the orm module.